creates new script to analyse statistics
